CS3247-26 Rework card nodes.

+++ Source/CS3247_Project/Cards/Crafting/Card Effects/Data/CardEffect.h
@@ -27,7 +27,7 @@ public:
 	UPROPERTY()
 	TMap<TSubclassOf<UAtomicCardEffect>, UAtomicCardEffect*> AtomicEffects;
 	
-	int32 EnchantmentDecay;
+	double EnchantmentDecay;
 	
 	UCardEffect() : AtomicEffects({}), EnchantmentDecay(0) {}
